+++ lld/docs/ReleaseNotes.rst
@@ -38,6 +38,9 @@ Breaking changes
 * ``-d`` is now ignored.
 * If a prevailing COMDAT group defines STB_WEAK symbol, having a STB_GLOBAL symbol in a non-prevailing group is now rejected with a diagnostic.
   (`D120626 <https://reviews.llvm.org/D120626>`_)
+* Support for the legacy ``.zdebug`` format has been removed. Run
+  ``objcopy --decompress-debug-sections`` in case old object files use ``.zdebug``.
+  (`D126793 <https://reviews.llvm.org/D126793>`_)
 
 COFF Improvements
 -----------------